+++ apps/builder/app/shared/copy-paste/use-copy-paste.ts
@@ -3,7 +3,7 @@ import { initCopyPaste } from "./init-copy-paste";
 import * as instance from "./plugin-instance";
 import * as embedTemplate from "./plugin-embed-template";
 import * as markdown from "./plugin-markdown";
-import * as webflow from "./plugin-webflow";
+import * as webflow from "./plugin-webflow/plugin-webflow";
 
 export const useCopyPaste = () => {
   useEffect(() => {
